requirements.txt
flask==2.3.3
flask-cors==4.0.0
rembg==2.0.50
Pillow==10.0.0
numpy==1.25.2 